+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Fixed usage data display for restriction categories
+* Rewritten migration process
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)